Custom fonts in QTextEdit, QTreeWidget and QCalendarWidget
Notes are supposed to be written in markdown and they will contain code
samples. Monospaced fonts are more suitable than proportional ones for this,
but proportional fonts can save space in tree views and a calendar.
Fonts will be divided into three groups: notes, tree views and a calendar.
Font and its size of every group will be customizable.

- Add font support to the stylesheet.
- Define all fonts and their sizes as constants - they will be defined
  in a config file late on.