Fix improper behaviour when another file is selected from a renamed note
Improper file is activated, when opened note is renamed. Code must be changed
to activate the file selected in the QTreeView. The problem is caused by
conflict from both signals (file changed and focus lost). Behaviour will be
changed. Notes will be saved when new note is selected or when the application
is closed. File will not be changed or application closed, if there is
conflict of file names - file with the same name is already present
in the directory.

Manual save will be implemented in the next commit.