FIX: Delete directory function crashing or behaving unexpectedly
There are several flaws. The whole process of deleting files will be
refactored. A sibling directory below the deleted one will be selected.
A sibling directory above the deleted one will be selected, if there is
no directory below it. A parent directory will be selected, if the deleted one
has no siblings. Input dialog requesting a name of the new directory
will appear, if the deleted directory was the last one in the notes
directory. A new empty note will be created, if the opened note is deleted.